AirQuality/requirements.txt
py-openaq==1.1.0
